requirements.txt
langchain
langchain_community
langchain_core
faiss-cpu
python-dotenv
streamlit

SpeechRecognition
pyaudio
pyttsx3
